+++ protected/humhub/docs/CHANGELOG.md
@@ -96,6 +96,7 @@ HumHub Change Log
 - Fix #2255: Added missing Social Account Settings menu
 - Fix: Added missing file download http caching
 - Enh: Added console email test command
+- Enh: Added stream module defaultStreamSuppressQueryIgnore to ease overwrites
 
 
 1.2.0-beta.2 (February 24, 2017)